+++ ubo_app/services/090-assistant/ubo-service/ubo_assistant/ubo_input_transport.py
@@ -1,6 +1,9 @@
 """Ubo Input Transport for Pipecat Reading Audio Samples from UBO RPC Client."""
 
+import asyncio
 import threading
+import time
+from dataclasses import dataclass
 from enum import StrEnum
 
 import numpy as np
@@ -19,9 +22,13 @@ from ubo_bindings.client import UboRPCClient
 from ubo_bindings.ubo.v1 import (
     Action,
     AudioReportSampleEvent,
+    AvailableInputDescription,
+    CameraReportImageEvent,
     DisplayRedrawAction,
     DisplayRenderEvent,
     Event,
+    InputDemandAction,
+    QrCodeInputDescription,
 )
 
 
@@ -30,6 +37,14 @@ class VideoSource(StrEnum):
 
     CAMERA = 'camera'
     DISPLAY = 'display'
+
+
+@dataclass
+class ReportedImage:
+    """Image frame structure for storing image data with a timestamp."""
+
+    timestamp: float
+    data: np.ndarray
 
 
 class UboInputTransport(BaseInputTransport):
@@ -47,9 +62,18 @@ class UboInputTransport(BaseInputTransport):
         self.audio_subscription = None
         self.audio_subscription_lock = threading.Lock()
 
-        self._image: dict[VideoSource, np.ndarray] = {
-            VideoSource.DISPLAY: np.zeros((0, 0, 3), dtype=np.uint8),
-            VideoSource.CAMERA: np.zeros((0, 0, 3), dtype=np.uint8),
+        self._image_set_events: dict[VideoSource, asyncio.Event] = {
+            key: asyncio.Event() for key in VideoSource
+        }
+        self._image: dict[VideoSource, ReportedImage] = {
+            VideoSource.DISPLAY: ReportedImage(
+                timestamp=0.0,
+                data=np.zeros((0, 0, 3), dtype=np.uint8),
+            ),
+            VideoSource.CAMERA: ReportedImage(
+                timestamp=0.0,
+                data=np.zeros((0, 0, 3), dtype=np.uint8),
+            ),
         }
 
         client.dispatch(
@@ -62,6 +86,10 @@ class UboInputTransport(BaseInputTransport):
             event_type=Event(display_render_event=DisplayRenderEvent()),
             callback=self._render_display,
         )
+        client.subscribe_event(
+            event_type=Event(camera_report_image_event=CameraReportImageEvent()),
+            callback=self._store_camera_image,
+        )
         super().__init__(params, **kwargs)
 
     async def process_frame(self, frame: Frame, direction: FrameDirection) -> None:
@@ -73,12 +101,35 @@ class UboInputTransport(BaseInputTransport):
             and frame.video_source in VideoSource.__members__.values()
         ):
             video_source = VideoSource(frame.video_source)
-            (h, w, *_) = self._image[video_source].shape
+
+            if video_source == VideoSource.CAMERA:
+                timestamp = time.time()
+                self.client.dispatch(
+                    action=Action(
+                        input_demand_action=InputDemandAction(
+                            description=AvailableInputDescription(
+                                qr_code_input_description=QrCodeInputDescription(
+                                    pattern=None,
+                                    title='Assistant Vision',
+                                    prompt=f'For prompt: {frame.context}',
+                                ),
+                            ),
+                        ),
+                    ),
+                )
+
+                event = self._image_set_events[video_source]
+                while self._image[video_source].timestamp < timestamp - 1:
+                    await event.wait()
+                    event.clear()
+
+            last_reported_image = self._image[video_source]
+            (h, w, *_) = last_reported_image.data.shape
 
             image_frame = UserImageRawFrame(
                 user_id='-',
                 request=frame,
-                image=self._image[video_source].tobytes(),
+                image=last_reported_image.data.tobytes(),
                 size=(w, h),
                 format='RGB',
             )
@@ -95,27 +146,40 @@ class UboInputTransport(BaseInputTransport):
             )
             data = render_event.data
             required_height = max(
-                self._image[VideoSource.DISPLAY].shape[0],
+                self._image[VideoSource.DISPLAY].data.shape[0],
                 canvas_width,
             )
             required_width = max(
-                self._image[VideoSource.DISPLAY].shape[1],
+                self._image[VideoSource.DISPLAY].data.shape[1],
                 canvas_height,
             )
             required_shape = (required_height, required_width, 3)
 
             if (
-                self._image[VideoSource.DISPLAY].shape[0] < required_height
-                or self._image[VideoSource.DISPLAY].shape[1] < required_width
+                self._image[VideoSource.DISPLAY].data.shape[0] < required_height
+                or self._image[VideoSource.DISPLAY].data.shape[1] < required_width
             ):
                 new_display = np.zeros(required_shape, dtype=np.uint8)
-                h, w = self._image[VideoSource.DISPLAY].shape[:2]
-                new_display[:h, :w, :] = self._image[VideoSource.DISPLAY]
-                self._image[VideoSource.DISPLAY] = new_display
-            self._image[VideoSource.DISPLAY][y1:y2, x1:x2, :] = np.frombuffer(
+                h, w = self._image[VideoSource.DISPLAY].data.shape[:2]
+                new_display[:h, :w, :] = self._image[VideoSource.DISPLAY].data
+                self._image[VideoSource.DISPLAY].data = new_display
+            self._image[VideoSource.DISPLAY].data[y1:y2, x1:x2, :] = np.frombuffer(
                 data,
                 dtype=np.uint8,
             ).reshape((y2 - y1, x2 - x1, 4))[:, :, :3]
+            self._image_set_events[VideoSource.DISPLAY].set()
+
+    def _store_camera_image(self, event: Event) -> None:
+        """Store the image coming from the camera."""
+        if image_event := event.camera_report_image_event:
+            self._image[VideoSource.CAMERA] = ReportedImage(
+                data=np.frombuffer(
+                    image_event.data,
+                    dtype=np.uint8,
+                ).reshape((image_event.height, image_event.width, 3)),
+                timestamp=time.time(),
+            )
+            self._image_set_events[VideoSource.CAMERA].set()
 
     def _set_is_listening(self, *, is_listening: bool) -> None:
         with self.audio_subscription_lock: